+++ src/dev/user.clj
@@ -1,5 +1,6 @@
 (ns user
   (:require
+    [clojure.java.io   :as io]
     [clojure.repl.deps :as crdeps]
     [clojure+.hashp    :as hashp]
     [clj-reload.core   :as reload]
@@ -18,17 +19,41 @@
 (defn reload! []
   (reload/reload))
 
+;; -----------------------------------------------------------------------------
+;; Printing helpers
+;; -----------------------------------------------------------------------------
+(defn- safe-requiring-resolve [sym]
+  (try
+    (deref (requiring-resolve sym))
+    (catch Exception _ nil)))
 
-(defn clear-terminal! []
-  (binding [*out* (java.io.PrintWriter. System/out)]
+
+(def original-out
+  (or
+    (:out (safe-requiring-resolve 'cider.nrepl.middleware.out/original-output))
+    System/out))
+
+
+(def original-err
+  (or
+    (:err (safe-requiring-resolve 'cider.nrepl.middleware.out/original-output))
+    System/err))
+
+
+(defmacro force-out
+  "Binds [[*out*]] to the original system's out."
+  [& body]
+  `(binding [*out* (io/writer original-out)]
+     ~@body))
+
+
+(defn clear-terminal!
+  "Clear the terminal of all text"
+  []
+  (force-out
     (print "\033c")
     (flush)))
 
-
-(defmacro force-out [& body]
-  `(binding [*out* (java.io.OutputStreamWriter. System/out)]
-     ~@body))
- 
 
 (comment
   (mdev/start! {:exception true})